@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 class Transaction {
   String id;
   String title;
@@ -6,9 +5,9 @@ class Transaction {
   DateTime date;
 
   Transaction({
-    @required this.id,
-    @required this.title,
-    @required this.amount,
-    @required this.date,
+    required this.id,
+    required this.title,
+    required this.amount,
+    required this.date,
   });
 }
