@@ -3,28 +3,31 @@ import 'package:flutter/cupertino.dart';
 import 'dart:io';
 
 class AdaptiveFlatButton extends StatelessWidget {
-  final String text;
-  final Function handler;
+  final String? text;
+  final Function()? handler;
 
   AdaptiveFlatButton(this.text, this.handler);
 
   @override
   Widget build(BuildContext context) {
     return Platform.isIOS
-                        ? CupertinoButton(
-                            child: Text(
-                              text,
-                              style: TextStyle(fontWeight: FontWeight.bold),
-                            ),
-                            onPressed: handler,
-                          )
-                        : FlatButton(
-                            textColor: Theme.of(context).primaryColor,
-                            child: Text(
-                              text,
-                              style: TextStyle(fontWeight: FontWeight.bold),
-                            ),
-                            onPressed: handler,
-                          );
+        ? CupertinoButton(
+            child: Text(
+              text!,
+              style: TextStyle(fontWeight: FontWeight.bold),
+            ),
+            onPressed: handler!,
+          )
+        : TextButton(
+            style: TextButton.styleFrom(
+                textStyle: TextStyle(
+              color: Theme.of(context).primaryColor,
+            )),
+            child: Text(
+              text!,
+              style: TextStyle(fontWeight: FontWeight.bold),
+            ),
+            onPressed: handler!,
+          );
   }
 }
