@@ -17,7 +17,7 @@ class TransactionList extends StatelessWidget {
                 children: <Widget>[
                   Text(
                     'No Transactions added yet!',
-                    style: Theme.of(context).textTheme.title,
+                    style: Theme.of(context).textTheme.headline6,
                   ),
                   SizedBox(
                     height: 30,
@@ -52,17 +52,20 @@ class TransactionList extends StatelessWidget {
                   ),
                   title: Text(
                     transactions[index].title,
-                    style: Theme.of(context).textTheme.title,
+                    style: Theme.of(context).textTheme.headline6,
                   ),
                   subtitle: Text(
                     DateFormat.yMMMd().format(transactions[index].date),
                   ),
                   trailing: MediaQuery.of(context).size.width > 360
-                      ? FlatButton.icon(
+                      ? TextButton.icon(
                           onPressed: () => deleteTx(transactions[index].id),
                           icon: Icon(Icons.delete),
                           label: Text('Delete'),
-                          textColor: Theme.of(context).errorColor,
+                          style: TextButton.styleFrom(
+                              textStyle: TextStyle(
+                            color: Theme.of(context).errorColor,
+                          )),
                         )
                       : IconButton(
                           icon: Icon(Icons.delete),
